+++ config/docker/server/start.sh
@@ -19,20 +19,22 @@ function catch()
     return $ex_code
 }
 
-name=$1
-mode=$2
-waitForOther=$3
-application="broker"
-theOtherApplication="server"
+catapultAppFolder=$1
+dataDirectory=$2
+application=$3
+theOtherApplication=$4
+mode=$5
+name=$6
+waitForOther=$7
 config="./$application-config"
 
 echo "RUNNING $application $name $mode"
-export LD_LIBRARY_PATH={{{catapultAppFolder}}}/lib:{{{catapultAppFolder}}}/deps
+export LD_LIBRARY_PATH=$catapultAppFolder/lib:$catapultAppFolder/deps
 
 ulimit -c unlimited
 
-rm -f "{{{dataDirectory}}}/$application.started"
-rm -f "{{{dataDirectory}}}/$application-recovery.started"
+rm -f "$dataDirectory/$application.started"
+rm -f "$dataDirectory/$application-recovery.started"
 
 if [ "$mode" == "DEBUG" ]; then
   echo "Setting up core dump..."
@@ -40,7 +42,7 @@ if [ "$mode" == "DEBUG" ]; then
   echo "./logs/$application.%e.%t" >/proc/sys/kernel/core_pattern
 fi
 
-otherApplicationRecoveryFile="{{{dataDirectory}}}/$theOtherApplication-recovery.started"
+otherApplicationRecoveryFile="$dataDirectory/$theOtherApplication-recovery.started"
 while [ -f $otherApplicationRecoveryFile ] ;
 do
     echo "Waiting for $theOtherApplication recovery to finish"
@@ -48,12 +50,12 @@ do
 done
 
 
-if [ -e "{{{dataDirectory}}}/$application.lock" ]; then
+if [ -e "$dataDirectory/$application.lock" ]; then
   echo "!!!! Have lock file present, going to run recovery in $application mode...."
 
-  touch "{{{dataDirectory}}}/$application-recovery.started"
+  touch "$dataDirectory/$application-recovery.started"
 
-  while [ -f "{{{dataDirectory}}}/$theOtherApplication.started" ] ;
+  while [ -f "$dataDirectory/$theOtherApplication.started" ] ;
   do
     echo "Waiting for $theOtherApplication to exit"
     sleep 1
@@ -62,31 +64,32 @@ if [ -e "{{{dataDirectory}}}/$application.lock" ]; then
   try
 (
     set -e
-    {{{catapultAppFolder}}}/bin/catapult.recovery "$config"
+    $catapultAppFolder/bin/catapult.recovery "$config"
     echo "!!!! Finished running recovery, should be moving on to start $application..."
 )
 # directly after closing the subshell you need to connect a group to the catch using ||
 catch || {
     echo "!!!! $application recovery has CRASHED!"
-    rm -f "{{{dataDirectory}}}/$application-recovery.started"
+    rm -f "$dataDirectory/$application-recovery.started"
     throw $ex_code # you can rethrow the "exception" causing the script to exit if not caught
 }
 fi
 
+rm -f "$dataDirectory/$application-recovery.started"
+
 if [ "$waitForOther" == "true" ]; then
-  while [ ! -f "{{{dataDirectory}}}/$theOtherApplication.started" ] ;
+  while [ ! -f "$dataDirectory/$theOtherApplication.started" ] ;
   do
   echo "Waiting for $theOtherApplication to start"
   sleep 1
   done
 fi
 
-rm -f "{{{dataDirectory}}}/$application-recovery.started"
-touch "{{{dataDirectory}}}/$application.started"
+touch "$dataDirectory/$application.started"
 
 processName="catapult.$application"
 echo "!!!! Starting $application...."
-{{{catapultAppFolder}}}/bin/$processName "$config" &
+$catapultAppFolder/bin/$processName "$config" &
 export applicationPid=$!
 
 cleanup() {
@@ -95,22 +98,21 @@ cleanup() {
   kill -s SIGTERM $applicationPid
   wait $applicationPid
   exitStatus=$?
-  rm "./data/$application.started"
+  rm -f "./data/$application.started"
   exit $exitStatus
 }
 
 #Trap SIGTERM
 trap 'cleanup' SIGTERM SIGINT
 
-
 while sleep 5; do
   ps aux |grep $processName |grep -q -v grep
   processStatus=$?
   if [ $processStatus -ne 0 ]; then
     echo "The process $processName has already exited."
-    wait $applicationPid 
+    wait $applicationPid
     exitStatus=$?
-    rm "./data/$application.started"
+    rm -f "./data/$application.started"
     exit $exitStatus
   fi
 
@@ -119,7 +121,7 @@ while sleep 5; do
     echo "kill -s SIGTERM $applicationPid"
     kill -s SIGTERM $applicationPid
     wait $applicationPid
-    rm "./data/$application.started"
+    rm -f "./data/$application.started"
     exit 1
   fi
 done